pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-benchmark"
version = "0.1.0"
description = "CLI benchmark harness for single-agent, multi-agent, and router-mode experiments."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=2.28.0",
  "openai-agents>=0.12.2",
  "pandas>=3.0.1",
  "pydantic>=2.12.5",
  "PyYAML>=6.0.3",
  "typer>=0.24.1",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.2"]
parquet = ["pyarrow>=18.0.0"]

[project.scripts]
benchmark = "agent_benchmark.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
